Day15 2_5D圖像產生器/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
onnxruntime==1.18.1
ultralytics>=8.3.0
openai>=1.40.0
sse-starlette==2.1.2
python-dotenv==1.0.1
imageio==2.34.1
